+++ cmd/influxd/run.go
@@ -470,6 +470,9 @@ func (cmd *RunCommand) openBroker(brokerURLs []url.URL, h *Handler) {
 
 	// Create broker
 	b := influxdb.NewBroker()
+	b.TruncationInterval = time.Duration(cmd.config.Broker.TruncationInterval)
+	b.MaxTopicSize = cmd.config.Broker.MaxTopicSize
+	b.MaxSegmentSize = cmd.config.Broker.MaxSegmentSize
 	cmd.node.Broker = b
 
 	// Create raft log.