+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 - [#2282](https://github.com/influxdb/influxdb/pull/2282): Use "value" as field name for OpenTSDB input.
 - [#2283](https://github.com/influxdb/influxdb/pull/2283): Fix bug when restarting an entire existing cluster.
 - [#2293](https://github.com/influxdb/influxdb/pull/2293): Open cluster listener before starting broker.
+- [#2287](https://github.com/influxdb/influxdb/pull/2287): Fix data race during SHOW RETENTION POLICIES.
+
+## Features
+- [#2276](https://github.com/influxdb/influxdb/pull/2276): Broker topic truncation.
 
 ## v0.9.0-rc24 [2015-04-13]